+++ doc-examples/src/main/java/react4j/examples/arez/step1/Footer.java
@@ -1,7 +1,7 @@
 package react4j.examples.arez.step1;
 
 import javax.annotation.Nullable;
-import react4j.annotations.EventHandler;
+import react4j.annotations.Callback;
 import react4j.annotations.ReactComponent;
 import react4j.arez.ReactArezComponent;
 import react4j.core.ReactNode;
@@ -18,7 +18,7 @@ import static react4j.examples.arez.step1.Footer_.*;
 abstract class Footer
   extends ReactArezComponent
 {
-  @EventHandler( MouseEventHandler.class )
+  @Callback( MouseEventHandler.class )
   void handleClearCompleted()
   {
     AppData.service.clearCompleted();