+++ processor/src/main/java/react4j/processor/Constants.java
@@ -7,7 +7,7 @@ final class Constants
   static final String DAGGER_PROVIDES_CLASSNAME = "dagger.Provides";
   static final String JS_FUNCTION_CLASSNAME = "jsinterop.annotations.JsFunction";
   static final String COMPONENT_CLASSNAME = "react4j.core.Component";
-  static final String EVENT_HANDLER_ANNOTATION_CLASSNAME = "react4j.annotations.EventHandler";
+  static final String CALLBACK_ANNOTATION_CLASSNAME = "react4j.annotations.Callback";
   static final String PROP_ANNOTATION_CLASSNAME = "react4j.annotations.Prop";
   static final String PROP_DEFAULT_ANNOTATION_CLASSNAME = "react4j.annotations.PropDefault";
   static final String REACT_COMPONENT_ANNOTATION_CLASSNAME = "react4j.annotations.ReactComponent";