+++ processor/src/test/resources/bad_input/com/example/callback/NotFunctionalInterface.java
@@ -1,25 +1,32 @@
-package com.example.prop;
+package com.example.callback;
 
-import javax.annotation.Nullable;
+import jsinterop.annotations.JsFunction;
 import react4j.annotations.Callback;
-import react4j.annotations.Prop;
 import react4j.annotations.ReactComponent;
 import react4j.core.BaseState;
 import react4j.core.Component;
 import react4j.core.ReactNode;
 
 @ReactComponent
-abstract class PropAndEventHandlerComponent
+abstract class NotFunctionalInterface
   extends Component<BaseState>
 {
-  @Prop
-  @Callback
-  protected abstract int getMyProp();
+  @JsFunction
+  interface Foo
+  {
+    void foo();
 
-  @Nullable
+    void foo2();
+  }
+
   @Override
   protected ReactNode render()
   {
     return null;
+  }
+
+  @Callback( Foo.class )
+  void handleFoo()
+  {
   }
 }