+++ processor/src/test/resources/bad_input/com/example/callback/Overload.java
@@ -1,0 +1,35 @@
+package com.example.callback;
+
+import jsinterop.annotations.JsFunction;
+import react4j.annotations.Callback;
+import react4j.annotations.ReactComponent;
+import react4j.core.BaseState;
+import react4j.core.Component;
+import react4j.core.ReactNode;
+
+@ReactComponent
+abstract class Overload
+  extends Component<BaseState>
+{
+  @JsFunction
+  public interface CustomHandler
+  {
+    void onMouseEvent( int i );
+  }
+
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+
+  @Callback( name = "a" )
+  void handleFoo()
+  {
+  }
+
+  @Callback( name = "b", value = CustomHandler.class )
+  void handleFoo( int i )
+  {
+  }
+}