+++ processor/src/test/resources/input/com/example/arez/ComponentWithNoAutoActionEventHandler.java
@@ -1,7 +1,7 @@
 package com.example.arez;
 
 import javax.annotation.Nullable;
-import react4j.annotations.EventHandler;
+import react4j.annotations.Callback;
 import react4j.annotations.ReactComponent;
 import react4j.arez.NoAutoAction;
 import react4j.arez.ReactArezComponent;
@@ -19,7 +19,7 @@ abstract class ComponentWithNoAutoActionEventHandler
   }
 
   @NoAutoAction
-  @EventHandler
+  @Callback
   void handleFoo()
   {
   }