+++ processor/src/test/resources/input/com/example/callback/CustomName.java
@@ -1,25 +1,23 @@
-package com.example.prop;
+package com.example.callback;
 
-import javax.annotation.Nullable;
 import react4j.annotations.Callback;
-import react4j.annotations.Prop;
 import react4j.annotations.ReactComponent;
 import react4j.core.BaseState;
 import react4j.core.Component;
 import react4j.core.ReactNode;
 
 @ReactComponent
-abstract class PropAndEventHandlerComponent
+abstract class CustomName
   extends Component<BaseState>
 {
-  @Prop
-  @Callback
-  protected abstract int getMyProp();
-
-  @Nullable
   @Override
   protected ReactNode render()
   {
     return null;
+  }
+
+  @Callback( name = "fox" )
+  void handleFoo()
+  {
   }
 }